Fix IndexedObjective bug with strings being removed.
Bug stems from the fact that Minecraft scoreboards accept unique entries only. When #put(int, String) is called with a String that already exists in the scoreboard it simply changes the score of the String entry, causing it to look like it is removing the old value.

The workaround is to keep appending RESET codes to the string entry until it is unique, if it reaches the max length (currently 40) it reverses and starts prepending RESET codes instead until it becomes a unique entry.

Voila